+++ Prac_02/CheckAdultOrNot.py
