+++ Prac_02/VolatileStock_byDay.py
