--- Prac_02_doneathome/ToPrint.txt
@@ -1,15 +1,0 @@
-$11.75
-$13.81
-$16.22
-$19.06
-$22.40
-$26.32
-$30.92
-$36.33
-$42.69
-$50.16
-$58.94
-$69.26
-$81.38
-$95.62
-$112.35
